+++ controllers/routers/account.js
@@ -38,7 +38,7 @@ exports.account.signin = {
         if (req.isAuthenticated()) return res.redirect('/home');
         return require('passport').authenticate('local', (err, authUser) => {
             if (err) return next(err);
-            if (!authUser) return next(new AccountError('Wrong email/username or password!'));
+            if (!authUser) return next(new AccountError(20002));
             return req.logIn(authUser, err => {
                 if (err) return next(err);
                 req.session.cookie.expires = req.body.isPersisted ? new Date(Date.now() + 14 * 24 * 3600000) : false;