+++ controllers/utilities/error.js
@@ -1,12 +1,17 @@
+const dictionaryProxy = require('./dictionary/');
+
+
 
 // ==============================
 //  CORE
 // ==============================
 class ExtendableError extends Error {
-    constructor(message) {
+    constructor(...arg) {
         if (new.target !== ExtendableError) {                                                                           // note: ensure this class cannot be instancified
-            super(message);
+            super();
             this.name = this.constructor.name;
+            this.message = dictionaryProxy[this.name](arg);                                                             // note: the proxy agent first try to map the arg with the dictionary of the error, if not found, return arg
+            if (typeof arg === 'number') this.code = arg[0];
             if (Error.captureStackTrace) Error.captureStackTrace(this, this.constructor);                               // note: only available on V8
             else this.stack = (new Error(message)).stack;                                                               // note: used if expose in client browser that not based on V8
         }
@@ -19,8 +24,8 @@ class ExtendableError extends Error {
 //  ERROR CLASSES
 // ==============================
 class ServerError       extends ExtendableError {}  // note: this error can not be handled by middleware
-
 class TemplateError     extends ExtendableError {}
+class HttpError         extends ExtendableError {}
 class AccountError      extends ExtendableError {   // tofix: mongo error handler redirecting
     constructor(arg) {
         if (arg instanceof ExtendableError) return arg;
@@ -30,18 +35,6 @@ class AccountError      extends ExtendableError {   // tofix: mongo error handle
             this.from = arg.name;
             this.stack = `AccountError (transferred):${new RegExp(/\s+at.+[^\n]/).exec(this.stack)[0]}\n${arg.stack}`;
         } else super(arg);
-    }
-}
-
-class HttpError         extends ExtendableError {
-    constructor(code) {
-        super();
-        this.code = code;
-        if (code === 400) this.message = 'HTTP 400 - Bad Request.';
-        if (code === 401) this.message = 'HTTP 401 - Unauthorized.';
-        if (code === 403) this.message = 'HTTP 403 - Forbidden.';
-        if (code === 404) this.message = 'HTTP 404 - Not found.';
-        if (code === 500) this.message = 'HTTP 500 - Internal Server Error.';
     }
 }